requirements/reqs.txt
fastapi==0.115.11
uvicorn==0.34.0
sqlmodel==0.0.24
psycopg2==2.9.10
alembic==1.13.1
passlib==1.7.4
ipython==9.1.0
pyjwt==2.10.1
